requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
cassandra-driver==3.24.0
click==7.1.2
Django==2.2
django-cassandra-engine==1.6.1
djangorestframework==3.12.2
geomet==0.2.1.post1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
